+++ AddOns/AutoPotion/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Auto Potion
 
-## [3.5.13](https://github.com/ollidiemaus/AutoPotion/tree/3.5.13) (2024-09-14)
-[Full Changelog](https://github.com/ollidiemaus/AutoPotion/compare/3.5.12...3.5.13) [Previous Releases](https://github.com/ollidiemaus/AutoPotion/releases)
+## [3.6.0](https://github.com/ollidiemaus/AutoPotion/tree/3.6.0) (2024-10-02)
+[Full Changelog](https://github.com/ollidiemaus/AutoPotion/compare/3.5.13...3.6.0) [Previous Releases](https://github.com/ollidiemaus/AutoPotion/releases)
 
-- cleanup  
+- fix reset button  
+- Cavedweller's Delight can now be configured